デザイン調整

+++ app/components/EntryList.tsx
@@ -5,11 +5,28 @@ import { useState, useEffect } from 'react';
 import { dateUtils } from '@/utils/date';
 import { getEntriesFromSheet, deleteEntryFromSheet, editEntryInSheet, getCategory } from '@/services/sheetService';
 import { Entry, Category } from '@/types/types';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import EditIcon from '@mui/icons-material/Edit';
+import ClearIcon from '@mui/icons-material/Clear';
+import CurrencyYenIcon from '@mui/icons-material/CurrencyYen';
+import StorefrontIcon from '@mui/icons-material/Storefront';
+import CategoryIcon from '@mui/icons-material/Category';
+import AssignmentIcon from '@mui/icons-material/Assignment';
+import Card from '@mui/material/Card';
+import CardContent from '@mui/material/CardContent';
+import Typography from '@mui/material/Typography';
+
 
 interface Props {
   entries: Entry[];
   categories: Category[];
 }
+
+const darkTheme = createTheme({
+  palette: {
+    mode: 'dark',
+  },
+});
 
 const EntryList = (props: Props) => {
     const [entries, setEntries] = useState<Entry[]>(props.entries);
@@ -27,7 +44,11 @@ const EntryList = (props: Props) => {
         setCategories(data);
     };
 
-    const onDelete = async (id: number) => {
+    const onDelete = async (e: React.MouseEvent<SVGSVGElement, MouseEvent>, id: number) => {
+        e.stopPropagation();
+        if (!confirm('削除しますか?')) {
+          return;
+        }
         await deleteEntryFromSheet(id);
         fetchEntries();
     };
@@ -43,35 +64,48 @@ const EntryList = (props: Props) => {
     };
 
     return (
-        <div className="container mx-auto p-4">
-            {editingEntry ? (
-                <>
-                    <AddEntryForm initialEntry={editingEntry} categories={categories} onSave={handleSave} />
-                    <button onClick={() => setEditingEntry(null)} className="mt-4 w-full flex justify-center py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-black bg-white hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
-                        キャンセル
-                    </button>
-                </>
-            ) : (
-                <ul className="max-w-2xl mx-auto p-4 bg-white rounded-lg shadow">
-                    {entries.map((entry, index) => (
-                        <li key={index} className="flex justify-between items-center p-4 border-b border-gray-200 last:border-0">
-                            <div className="flex-grow">
-                                <div className="text-sm text-gray-600">{dateUtils.formatDateToDateTimeLocal(entry.date)}</div>
-                                <div className="text-lg font-medium">{entry.category ? entry.category.name : 'No Category'}</div>
-                                <div className="text-gray-800">{entry.amount} 円</div>
-                                <div className="text-gray-500 italic">{entry.store}</div>
-                                <div className="text-gray-500 italic">{entry.memo}</div>
-                            </div>
-                            <div className="flex items-center space-x-2">
-                                <button onClick={() => onEdit(entry)} className="px-3 py-1 bg-blue-600 text-white rounded hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2">編集</button>
-                                <button onClick={() => onDelete(entry.id)} className="px-3 py-1 bg-red-600 text-white rounded hover:bg-red-700 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2">削除</button>
-                            </div>
-                        </li>
-                    ))}
-                </ul>
-            )}
-        </div>
+        <ThemeProvider theme={darkTheme}>
+          <div className="p-4">
+              {editingEntry ? (
+                  <>
+                      <AddEntryForm initialEntry={editingEntry} categories={categories} onSave={handleSave} />
+                      <button onClick={() => setEditingEntry(null)} className="mt-4 w-full flex justify-center py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-black bg-white hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+                          キャンセル
+                      </button>
+                  </>
+              ) : (
+                <div>
+                  {entries.map((entry, index) => (
+                    <div key="entry.id">
+                      {(index === 0 || dateUtils.formatDate(entry.date) !== dateUtils.formatDate(entries[index - 1].date)) && (
+                        <div className="mb-4">
+                          <Typography variant="subtitle1" component="strong" color="white">
+                            {dateUtils.formatDate(entry.date)} ({dateUtils.getDayOfWeek(entry.date)})
+                          </Typography>
+                        </div>
+                      )}
 
+                      <Card className="mb-4">
+                        <CardContent sx={{ cursor: 'pointer' }} onClick={() => onEdit(entry)}>
+                          <div>
+                              <div className="flex items-center justify-between text-lg">
+                                <div className="flex items-center"><CurrencyYenIcon/>{entry.amount}</div>
+                                <ClearIcon className="cursor-pointer" onClick={(e) => onDelete(e, entry.id)}/>
+                              </div>
+                              <div className="flex items-center">
+                                <CategoryIcon/>{entry.category ? entry.category.name : 'No Category'}
+                              </div>
+                              <div className="flex items-center"><StorefrontIcon/>{entry.store}</div>
+                              <div className="flex items-center"><AssignmentIcon/>{entry.memo}</div>
+                          </div>
+                        </CardContent>
+                      </Card>
+                    </div>
+                  ))}
+                </div>
+              )}
+          </div>
+        </ThemeProvider>
     );
 };
